rename(move): 파일 패키지 이동 - 상담 프래그먼트 view 패키지로 이동
