+++ app/src/main/res/layout/fragment_counseling_daily_praise.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.counseling.CounselingDailyPraiseFragment">
+    tools:context=".ui.counseling.view.CounselingDailyPraiseFragment">
 
     <androidx.appcompat.widget.LinearLayoutCompat
         android:id="@+id/ll_counseling_daily_praise_notification"
