+++ app/src/main/res/layout/fragment_counseling_statistics.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.counseling.CounselingStatisticsFragment">
+    tools:context=".ui.counseling.view.CounselingStatisticsFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
