+++ app/src/main/res/layout/fragment_counseling_weekly_report.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.counseling.CounselingWeeklyReportFragment">
+    tools:context=".ui.counseling.view.CounselingWeeklyReportFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
